EES-1548 Update content API endpoints to better follow REST principles
This change includes:
- Removing unnecessary `/content` prefix from various routes.
- General improvements to URLs to make them more RESTful.
- Adding new `ThemeController` which responds with the 'tree' view
  models. We have separate endpoints for the respective theme types -
  downloads, methodologies and publications.
