+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api.Tests/Controllers/MethodologyControllerTests.cs
@@ -12,71 +12,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Tests.Controlle
     public class MethodologyControllerTests
     {
         [Fact]
-        public void Get_MethodologyTree_Returns_Ok()
-        {
-            var fileStorageService = new Mock<IBlobStorageService>();
-
-            fileStorageService.Setup(
-                s => s.DownloadBlobText(
-                    PublicContentContainerName,
-                    "methodology/tree.json"
-                )
-            ).ReturnsAsync(
-                @"
-            [
-            {
-                ""id"": ""5dde1ee6-34b6-41f5-bdc0-10f7fe0c46fe"",
-                ""title"": ""string"",
-                ""summary"": ""string"",
-                ""topics"": [
-                {
-                    ""id"": ""46dae1b0-b2ec-4759-b04e-84d184afe6db"",
-                    ""title"": ""string"",
-                    ""summary"": ""string"",
-                    ""publications"": [
-                    {
-                        ""methodology"": {
-                            ""id"": ""c3fc42f6-996a-4d8b-9b8d-98a536197ea7"",
-                            ""slug"": ""string"",
-                            ""summary"": ""string"",
-                            ""title"": ""string""
-                        },
-                        ""id"": ""c00e6006-e82e-409d-920e-51860df6e076"",
-                        ""title"": ""string"",
-                        ""slug"": ""string"",
-                        ""summary"": ""string""
-                    }
-                    ]
-                }
-                ]
-            }
-            ]"
-            );
-
-            var controller = new MethodologyController(fileStorageService.Object);
-
-            var result = controller.GetMethodologyTree();
-            Assert.Single(result.Result.Value);
-            var theme = result.Result.Value.First();
-            Assert.IsAssignableFrom<ThemeTree<MethodologyTreeNode>>(theme);
-            Assert.Single(theme.Topics);
-            var topic = theme.Topics.First();
-            Assert.Single(topic.Publications);
-        }
-
-        [Fact]
-        public void Get_MethodologyTree_Returns_NoContent()
-        {
-            var fileStorageService = new Mock<IBlobStorageService>();
-
-            var controller = new MethodologyController(fileStorageService.Object);
-
-            var result = controller.GetMethodologyTree();
-
-            Assert.IsAssignableFrom<NoContentResult>(result.Result.Result);
-        }
-
-        [Fact]
         public void Get_Methodology_Returns_Ok()
         {
             var fileStorageService = new Mock<IBlobStorageService>();
